Move login, logout responses.
