+++ packages/admin/routes/web.php
@@ -2,7 +2,7 @@
 
 use Filament\Facades\Filament;
 use Filament\Http\Controllers\AssetController;
-use Filament\Http\Responses\Auth\Contracts\LogoutResponse;
+use Filament\Contracts\Authentication\LogoutResponse;
 use Illuminate\Support\Facades\Route;
 
 Route::domain(config('filament.domain'))
