+++ packages/admin/src/Http/Responses/Auth/LoginResponse.php
@@ -3,7 +3,7 @@
 namespace Filament\Http\Responses\Auth;
 
 use Filament\Facades\Filament;
-use Filament\Http\Responses\Auth\Contracts\LoginResponse as Responsable;
+use Filament\Contracts\Authentication\LoginResponse as Responsable;
 use Illuminate\Http\RedirectResponse;
 use Livewire\Redirector;
