+++ packages/admin/src/Http/Responses/Auth/LogoutResponse.php
@@ -2,7 +2,7 @@
 
 namespace Filament\Http\Responses\Auth;
 
-use Filament\Http\Responses\Auth\Contracts\LogoutResponse as Responsable;
+use Filament\Contracts\Authentication\LogoutResponse as Responsable;
 use Illuminate\Http\RedirectResponse;
 
 class LogoutResponse implements Responsable
